get pagination products doc

+++ src/dao/mongoManagers/productManager.js
@@ -1,4 +1,6 @@
 import { productsModel } from "../models/products.model.js";
+import CustomError from "../../errors/CustomError.js";
+import { ErrorsName, ErrorsCause, ErrorsMessage } from "../../errors/errors.enum.js";
 
 export default class ProductManager {
     async getProducts() {
@@ -13,7 +15,6 @@ export default class ProductManager {
     async getPagination(category, page, limit) {
         try {
             const products = await productsModel.paginate( category , page, limit );
-            console.log(products);
             return products;
         } catch (error) {
             console.log(error);
@@ -22,19 +23,35 @@ export default class ProductManager {
 
     async getProductById(id) {
         try {
-            const products = await productsModel.findById(id);
-            return products;
+            if (id.length === 24) {
+                const products = await productsModel.findById(id);
+                return products;
+            } else {
+                CustomError.createCustomError({
+                    name: ErrorsName.PRODUCT_DATA_INCOMPLETE,
+                    message: ErrorsMessage.PRODUCT_DATA_INCOMPLETE,
+                    cause: ErrorsCause.PRODUCT_DATA_INCOMPLETE
+                });
+            }
         } catch (error) {
-            console.log(error);
+            throw error;
         }
     }
 
     async addProduct(obj) {
         try {
-            const newProduct = await productsModel.create(obj);
-            return newProduct;
+            if (obj) {
+                const newProduct = await productsModel.create(obj);
+                return newProduct;
+            } else {
+                CustomError.createCustomError({
+                    name: ErrorsName.PRODUCT_DATA_INCOMPLETE,
+                    message: ErrorsMessage.PRODUCT_DATA_INCOMPLETE,
+                    cause: ErrorsCause.PRODUCT_DATA_INCOMPLETE
+                });
+            }
         } catch (error) {
-            console.log(error);
+            throw error;
         }
     }
 
@@ -43,16 +60,24 @@ export default class ProductManager {
             const updateProduct = await productsModel.findByIdAndUpdate(id, obj);
             return updateProduct;
         } catch (error) {
-            console.log(error);
+            throw error;
         }
     }
 
     async deleteProduct(id) {
         try {
-            const deletedProduct = await productsModel.findByIdAndDelete(id);
-            return deletedProduct;
+            if (id.length === 24) {
+                const deletedProduct = await productsModel.findByIdAndDelete(id);
+                return deletedProduct;
+            } else {
+                CustomError.createCustomError({
+                    name: ErrorsName.PRODUCT_DATA_INCOMPLETE,
+                    message: ErrorsMessage.PRODUCT_DATA_INCOMPLETE,
+                    cause: ErrorsCause.PRODUCT_DATA_INCOMPLETE
+                });
+            }
         } catch (error) {
-            console.log(error);
+            throw error;
         }
     }
 }
